LayoutTests/http/tests/storageAccess/grant-storage-access-under-opener-expected.txt

CONSOLE MESSAGE: line 8: First JavaScript statement.
CONSOLE MESSAGE: line 57: Running test.
CONSOLE MESSAGE: line 73: About to call testRunner.statisticsUpdateCookiePartitioning().
CONSOLE MESSAGE: line 75: About to open the new window.
CONSOLE MESSAGE: line 45: About to open the third party iframe.
Tests that a cross-origin window from a prevalent domain with non-recent user interaction gets immediate storage access under its opener.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Cookie created.
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
Should receive first-party cookie.
Received cookie named 'firstPartyCookie'.
Did not receive cookie named ''.
Client-side document.cookie: firstPartyCookie=value
